1.1.0 some previous builds Develop

+++ src/main/java/elucent/rootsclassic/component/ComponentBase.java
@@ -13,13 +13,11 @@ import net.minecraft.world.World;
 public class ComponentBase {
 
   private String name = "";
-
   private ItemStack itemSource = null;
   public Vec3d primaryColor = new Vec3d(0, 0, 0);
   public Vec3d secondaryColor = new Vec3d(0, 0, 0);
-  private int manaCost = 0;
-
- private  TextFormatting textColor = TextFormatting.WHITE;
+  private float manaCost = 0;
+  private TextFormatting textColor = TextFormatting.WHITE;
 
   public ComponentBase setPrimaryColor(double r, double g, double b) {
     this.primaryColor = new Vec3d(r, g, b);
@@ -80,11 +78,11 @@ public class ComponentBase {
 
   public void castingAction(EntityPlayer player, int count, int potency, int efficiency, int size) {}
 
-  public int getManaCost() {
+  public float getManaCost() {
     return manaCost;
   }
 
-  public void setManaCost(int manaCost) {
+  public void setManaCost(float manaCost) {
     this.manaCost = manaCost;
   }
 }
